fix: extract PyVistaBackend to viewer_backends_pyvista.py to fix LOC budget
viewer_backends.py exceeded the 1200 LOC CI hard limit (1217 lines).
Extracted PyVistaBackend (253 LOC) to a dedicated module and replaced the
direct reference in create_viewer() with a lazy import, reducing the main
file to 966 LOC.
